Pull the IP address of the NameNode on the host which is running the container from consul
refs CIO-33866
refs CIO-33868

Each HDFS protocol endpoint listening on the tenant network is registered
with consul as the service with name 'namenode', tagged with the host's
node-uuid.

On container startup, we now query for the IP of the HDFS protocol endpoint
on the host running the container and add it as an entry to /etc/hosts as
the host named 'namenode'.

Also, remove some content that is no longer used:
* local consul agent binaries and configuration files
* /usr/local/bin/cio-hadoop-topology script
* /usr/bin/do-with-consul entrypoint script
* openssh-server
* jq